% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episphereR.R
\name{boxRead}
\alias{boxRead}
\title{boxRead}
\usage{
boxRead(file_id = 787813783402, version_id = NULL, ...)
}
\arguments{
\item{file_id}{the file to read (default is a simulated data file)}

\item{version_id}{the version of the box file}

\item{...}{values passed to boxr::box_read}
}
\value{
an R object
}
\description{
read data from a box file.
The function is a wrapper around the box_read function
}
\examples{
boxRead(file_id=787813783402)
}
\seealso{
\code{\link[boxr]{box_read}}
}
