% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{lookup_concept}
\alias{lookup_concept}
\title{lookup_concept - search concept dictionary for terms}
\usage{
lookup_concept(
  x,
  dict = "https://episphere.github.io/conceptGithubActions/aggregate.json"
)
}
\arguments{
\item{x}{The word/substring that you want to find.}

\item{dict}{The JSON dictionary

Looks up the word/substring in JSON dictionary.  The format of the
dictionary is as follows.
{
  "concept_id":{
      "label":variable label
      "name": variable name
  }
}

 See \href{https://episphere.github.io/conceptGithubActions/aggregate.json}{this example}
 which is the default dictionary.}
}
\description{
Looks in the variable label/name for a string/substring
}
