% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episphereR.R
\name{openBoxFile}
\alias{openBoxFile}
\title{opens a box file in your default browser}
\usage{
openBoxFile(id = 787813783402, ver = NULL)
}
\arguments{
\item{id}{box file id.}

\item{ver}{box file version number}
}
\value{
opens a browser window (returns NULL)
}
\description{
opens a box file in your default browser
}
\examples{
openBoxFile()
openBoxFile(id=787813783402)
}
