% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episphereR.R
\name{openBoxFolder}
\alias{openBoxFolder}
\title{opens the box file browser for a directory in your default browser}
\usage{
openBoxFolder(id = 133596945131)
}
\arguments{
\item{id}{box folder id.}
}
\value{
opens a browser window (returns NULL)
}
\description{
opens the box file browser for a directory in your default browser
}
\examples{
openBoxFolder()
openBoxFolder(id=0)
}
