% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episphereR.R
\name{openURL}
\alias{openURL}
\title{open a browser window}
\usage{
openURL(url = "https://episphere.github.io/r")
}
\arguments{
\item{url}{that the browser opens.}
}
\value{
opens a browser window (returns NULL)
}
\description{
open a browser window
}
\examples{
openURL()
openURL("https://www.google.com")
}
